docs: 📝 Add usage instructions
What our documentation was still missing were details about how to get
the bot to join a Discord server (ie how to generate a Discord Token)
and what was needed on the Gitea side. Also we previously ommited
how to interact with the bot once it was joined to a server.

What we're still missing, are details on the Discord Token Scopes that
are required.

fixes #6